feat: Replace HorizontalScrollBox with ScrollBox component for improved flexibility and scrolling behavior
