--- src/component/game/common/layout/style/HorizontalScrollBox.module.css
@@ -1,9 +1,0 @@
-.scrollBox {
-  display: flex;
-  overflow-x: auto;
-  white-space: nowrap;
-  scroll-behavior: smooth;
-  box-sizing: border-box;
-  padding-bottom: 5px;
-  gap: 5px;
-}
